+++ cmd/data-node/node/node.go
@@ -108,7 +108,8 @@ type NodeCommand struct {
 
 	vegaCoreServiceClient vegaprotoapi.CoreServiceClient
 
-	broker *broker.Broker
+	broker    *broker.Broker
+	sqlBroker broker.SqlStoreEventBroker
 
 	transferRespSub      *subscribers.TransferResponse
 	marketEventSub       *subscribers.MarketEvent
@@ -258,6 +259,12 @@ func (l *NodeCommand) runNode(args []string) error {
 	eg.Go(func() error {
 		return l.broker.Receive(ctx)
 	})
+
+	if l.conf.SQLStore.Enabled {
+		eg.Go(func() error {
+			return l.sqlBroker.Receive(ctx)
+		})
+	}
 
 	// waitSig will wait for a sigterm or sigint interrupt.
 	eg.Go(func() error {